[UPD] Refactorisation.
