+++ aliments-angular-capacitor-fields-relational/src/app/modals/aliment-modals/aliment-modals.component.ts
@@ -8,7 +8,7 @@ import {
 import { NgbModal, NgbModalRef } from '@ng-bootstrap/ng-bootstrap';
 import { ErrorHandlerService } from 'src/app/services/error-handler.service';
 import { AlimentService } from 'src/app/services/aliment.service';
-import { ErplibreRestAlimentService } from 'src/app/services/erplibre-rest-aliment.service';
+import { ErplibreRestAlimentService } from 'src/app/services/erplibre-rest/erplibre-rest-aliment.service';
 import { ModalOpenerService } from 'src/app/services/modal-opener.service';
 import { AlimentModel } from 'src/models/aliment.model';
