+++ aliments-angular-capacitor-fields-relational/src/app/modals/recipe-modals/recipe-modals.component.scss
@@ -1,0 +1,4 @@
+.recipe-modal--info__field:empty::after {
+	content: ".";
+	visibility: hidden;
+}
