+++ aliments-angular-capacitor-fields-relational/src/app/services/aliment.service.ts
@@ -1,7 +1,7 @@
 import { Injectable } from '@angular/core';
 import { Observable, Subject } from 'rxjs';
 import { AlimentModel } from 'src/models/aliment.model';
-import { ErplibreRestAlimentService } from './erplibre-rest-aliment.service';
+import { ErplibreRestAlimentService } from './erplibre-rest/erplibre-rest-aliment.service';
 
 @Injectable({
 	providedIn: 'root',
@@ -15,14 +15,9 @@ export class AlimentService {
 	}
 
 	constructor(private erplibreRest: ErplibreRestAlimentService) {
-		this.erplibreRest.getAliments().subscribe({
-			next: (getResponse) => {
-				this._aliments = getResponse;
-				this.aliments.next(this._aliments);
-			},
-			error: (e) => {
-				console.error(e);
-			},
+		this.erplibreRest.getAliments().subscribe((getResponse) => {
+			this._aliments = getResponse;
+			this.aliments.next(this._aliments);
 		});
 	}
